requirements.txt
Django==1.10.1
Markdown==2.6.7
Pillow==3.3.1
dj-database-url==0.4.1
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
whitenoise==3.2.2
psycopg2==2.6.2
